realtion link with User model and Video model, some authorization works

+++ src/controllers/videoController.js
@@ -1,4 +1,6 @@
 import Video from "../models/Video"; 
+import User from "../models/User";
+import { compareSync } from "bcrypt";
 
 
 export const home = async (req, res) => {
@@ -9,7 +11,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video){
     return res.render("404", { pageTitle: "Video not found." });
   }
@@ -18,21 +20,33 @@ export const watch = async (req, res) => {
 
 
 export const getEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const video = await Video.findById(id);
   if (!video){
     return res.render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit-video", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({_id: id});
   if (!video){
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -44,7 +58,17 @@ export const postEdit = async (req, res) => {
 
 
 export const deleteVideo = async (req, res) => {
+  const {
+    user: { _id }
+  } = req.session;
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/")
 };
@@ -56,16 +80,23 @@ export const getUpload = (req, res) => {
 
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { path:fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
-      hashtags,
+      hashtags: Video.formatHashtags(hashtags),
       fileUrl,
+      owner: _id,
       createdAt: Date.now(),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {pageTitle: "Upload Video", errorMessage: error._message,});
